마이페이지 내 즐겨찾기, 리뷰 로직 작성/레이아웃 디자인 필요

+++ app/src/main/java/com/cookandroid/muksaegwon/model/StoreReview.java
@@ -3,9 +3,9 @@ package com.cookandroid.muksaegwon.model;
 public class StoreReview {
     private String review;
     private String date;
-    private short rating;
+    private float rating;
 
-    public StoreReview(String review, String date, short rating) {
+    public StoreReview(String review, String date, float rating) {
         this.review = review;
         this.date = date;
         this.rating = rating;
@@ -27,7 +27,7 @@ public class StoreReview {
         this.date = date;
     }
 
-    public short getRating() {
+    public float getRating() {
         return rating;
     }
